requirements.txt
python-dotenv==1.0.1
langchain==0.2.15
langchain-community==0.2.12
openai==1.43.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
psycopg[binary]==3.2.1
pandas==2.2.2
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
